Add first API for notebook model

+++ jupyter_nbmodel_client/constants.py
@@ -3,5 +3,5 @@ import re
 
 HTTP_PROTOCOL_REGEXP = re.compile(r"^http")
 """http protocol regular expression."""
-REQUEST_TIMEOUT = int(os.getenv("REQUEST_TIMEOUT", 10))
+REQUEST_TIMEOUT = float(os.getenv("REQUEST_TIMEOUT", 10))
 """Default request timeout in seconds"""
